+++ permissions/tests/base.py
@@ -7,15 +7,34 @@ from permissions import PermissionsRegistry as BasePermissionsRegistry
 
 class PermissionsRegistry(BasePermissionsRegistry):
 
+    def _get_user_model(self):
+        return User
+
     def _get_model_instance(self, model, **kwargs):
         return model(**kwargs)
 
 
-class Model:
+class Model(object):
 
     def __init__(self, **kwargs):
         for k, v in kwargs.items():
             setattr(self, k, v)
+
+
+class User(Model):
+
+    def __init__(self, **kwargs):
+        kwargs.setdefault('permissions', [])
+        super(User, self).__init__(**kwargs)
+
+    def is_anonymous(self):
+        return False
+
+
+class AnonymousUser(User):
+
+    def is_anonymous(self):
+        return True
 
 
 class TestCase(BaseTestCase):