+++ CHANGELOG.md
@@ -3,13 +3,24 @@
 ## [Unreleased] - [unreleased]
 
 ### Deprecated
+
 - Object-notation access to permissions will be removed in 2.0. All views
   should be decorated using `@permissions.require('perm_name')` instead of
   `@permissions.perm_name`.
 
+### Changed
+
+- Template filters now immediately return False if they're passed a non-user
+  object (that is, an object that is not an instance of get_user_model() or
+  AnonymousUser). This makes these template filters less likely to crash the
+  page.
+- Template filters now bypass the call to the permission function for anonymous
+  users when anonymous users aren't allowed.
+
 ## [1.1.0] - 2015-03-27
 
 ### Fixed
+
 - Template filters now bypass the call to the permission function when
   appropriate (e.g., when `allow_staff=True` and the user is staff). The
   logic for this was already in place but was broken due to operator
@@ -19,6 +30,7 @@
   `MIDDLEWARE_CLASSES` setting).
 
 ### Added
+
 - `PermissionsRegistry.entry_for_view(view, permission_name)` can be used to
   whether a view requires a permission. This is intended for use in testing.
 - Travis CI config was added.